requirements.txt
# ============================================================
# 统一数据管道依赖
# 涵盖 RAG 管道 + 分子提取管道的所有依赖
# ============================================================

# ──────────────────────────────────────────────────────────────
# PDF 转 Markdown（共享模块 pdf_converter.py）
# ──────────────────────────────────────────────────────────────
docling>=2.0.0
docling-core>=2.0.0

# ──────────────────────────────────────────────────────────────
# LLM / OpenAI
# ──────────────────────────────────────────────────────────────
openai>=1.0.0
anthropic>=0.20.0

# ──────────────────────────────────────────────────────────────
# RAG 管道（rag/ 目录）
# ──────────────────────────────────────────────────────────────
# 步骤 2: Pinecone 向量化
pinecone>=3.0.0
langchain-text-splitters>=0.0.1

# 步骤 3: S3 同步
boto3>=1.26.0

# ──────────────────────────────────────────────────────────────
# 分子提取管道（molecule/ 目录）
# ──────────────────────────────────────────────────────────────
# 步骤 2: Snowflake 元数据更新
snowflake-connector-python>=3.0.0

# llm_judge: 分子处理
rdkit>=2023.3.1
pubchempy>=1.0.4
tiktoken>=0.5.0
tavily-python>=0.3.0
ollama>=0.1.0
pydantic>=2.0.0

# ──────────────────────────────────────────────────────────────
# 数据处理
# ──────────────────────────────────────────────────────────────
pandas>=1.5.0
numpy>=1.24.0
openpyxl>=3.0.0
matplotlib>=3.6.0

# ──────────────────────────────────────────────────────────────
# 网络与 API 调用
# ──────────────────────────────────────────────────────────────
requests>=2.28.0
tenacity>=8.0.0

# ──────────────────────────────────────────────────────────────
# 工具库
# ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.64.0
python-dateutil>=2.8.0
setuptools>=65.0.0
